# clients/go/api/openapi.yaml
openapi: 3.0.0
info:
  description: "API REST gamificada para gerenciamento de tarefas com IA, XP, levels\
    \ e sistema de recompensas"
  title: Gamified Task Management API
  version: 1.0.0
servers:
- description: Servidor de Desenvolvimento
  url: http://localhost:3000
security:
- bearerAuth: []
...
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      parameters:
      - description: Filtrar por status
        explode: true
        in: query
        name: status
        required: false
        schema:
          enum:
          - pending
          - in_progress
          - completed
          type: string
        style: form
      - description: Filtrar por prioridade
        explode: true
        in: query
        name: priority
        required: false
        schema:
          enum:
          - low
          - medium
          - high
          type: string
        style: form
      - description: Número da página
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Itens por página
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
        style: form
      - description: Campo para ordenação
        explode: true
        in: query
        name: sortBy
        required: false
        schema:
          default: createdAt
          enum:
          - createdAt
          - updatedAt
          - dueDate
          - priority
          - title
          type: string
        style: form
      - description: Ordem da classificação
        explode: true
        in: query
        name: sortOrder
        required: false
        schema:
          default: desc
          enum:
          - asc
          - desc
          type: string
        style: form
      - description: Buscar no título e descrição
        explode: true
        in: query
        name: search
        required: false
        schema:
          maxLength: 100
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listTasks_200_response'
          description: Lista de tarefas
      security:
      - bearerAuth: []
      summary: Get all tasks for the logged-in user
      tags:
      - Tasks
    post:
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createTask_request'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createTask_201_response'
          description: Tarefa criada com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Dados inválidos
      security:
      - bearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /api/tasks/stats:
    get:
      operationId: getTaskStats
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTaskStats_200_response'
          description: Estatísticas das tarefas
      security:
      - bearerAuth: []
      summary: Get task statistics
      tags:
      - Tasks
  /api/tasks/status/{status}:
    get:
      operationId: listTasksByStatus
      parameters:
      - description: Status das tarefas
        explode: false
        in: path
        name: status
        required: true
        schema:
          enum:
          - pending
          - in_progress
          - completed
          type: string
        style: simple
      - description: Número da página
        explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - description: Itens por página
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listTasksByStatus_200_response'
          description: Tarefas filtradas por status
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Status inválido
      security:
      - bearerAuth: []
      summary: Get tasks by status
      tags:
      - Tasks
  /api/tasks/{id}:
    delete:
      operationId: deleteTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/deleteTask_200_response'
          description: Tarefa deletada com sucesso
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      operationId: getTaskById
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTaskById_200_response'
          description: Detalhes da tarefa
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Get a specific task by ID
      tags:
      - Tasks
    put:
      operationId: updateTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTask_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateTask_200_response'
          description: Tarefa atualizada com sucesso
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Dados inválidos
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Update a task
      tags:
      - Tasks
  /api/tasks/{id}/complete:
    patch:
      operationId: completeTask
      parameters:
      - description: ID da tarefa
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/completeTask_200_response'
          description: Tarefa marcada como concluída
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa já está concluída
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Acesso negado
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Tarefa não encontrada
      security:
      - bearerAuth: []
      summary: Mark a task as complete
      tags:
      - Tasks
  /api/tasks/story:
    post:
      operationId: createStoryFromTasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createStoryFromTasks_request'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createStoryFromTasks_201_response'
          description: História épica gerada com sucesso
      security:
      - bearerAuth: []
      summary: Generate a story from tasks
      tags:
      - Gamification
  /api/tasks/{id}/complete-quest:
    post:
      operationId: completeQuest
      parameters:
      - description: ID da tarefa/quest
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/completeQuest_request'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/completeQuest_200_response'
          description: Quest completada com sucesso
      security:
      - bearerAuth: []
      summary: Complete a quest (task) and get XP
      tags:
      - Gamification
...
components:
  schemas:
...
    listTasks_200_response:
      example:
        data:
          pagination:
            current: 1
            total: 50
            pages: 5
            limit: 10
          tasks:
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
          - id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
      properties:
        success:
          example: true
          type: boolean
        data:
          $ref: '#/components/schemas/listTasks_200_response_data'
      type: object
...
    createTask_201_response:
      example:
        data:
          task:
            id: 64a1b2c3d4e5f6789abcdef1
            title: Implementar Story-Mode
            priority: low
            userId: 64a1b2c3d4e5f6789abcdef0
            status: pending
        success: true
        message: Tarefa criada com sucesso
      properties:
        success:
          example: true
          type: boolean
        message:
          example: Tarefa criada com sucesso
          type: string
        data:
          $ref: '#/components/schemas/createTask_201_response_data'
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http